+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 **Fixed bugs:**
 
-- Activated Function `New-VirtualEnvLocal` for downloading packages of virtual environment, e.g. for offline installation.
+- Activated function `New-VirtualEnvLocal` for downloading packages of virtual environment, e.g. for offline installation.
 - Added opportunity for offline installation of a folder with local packages
 
 ## [0.5.3](https://github.com/wbrandenburger/PSVirtualEnv/tree/0.5.3) (2019-09-17)