+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 **Fixed bugs:**
 
 - Activated function `New-VirtualEnvLocal` for downloading packages of virtual environment, e.g. for offline installation.
-- Added opportunity for offline installation of a folder with local packages
+- Added opportunity for offline installation of a folder with local packages.
 
 ## [0.5.3](https://github.com/wbrandenburger/PSVirtualEnv/tree/0.5.3) (2019-09-17)